updated readme
updated the Google assistant fan control language to make it clear how it behaves. Great work on this Repo btw. (I know this is not the right place, but please consider adding in an update to the preference value for dimmerControl of the "King of Fans Zigbee Fan Controller" to have a value of "Fan". This will fix Alexa integrations for others since that is what I had to do to get it to work. Before, the slider control for both fan and light only controls the light in Alexa. This also does not break the smartthings controller or the child device (light) just as an FYI.)